(refactor) move cache_files from radis/misc to radis/io
